requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==24.1
plotly==5.24.1
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
retrying==1.3.4
selenium==4.22.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
zipp==3.20.2

psycopg2~=2.9.9
twilio~=9.3.3
python-dotenv~=1.0.1